refactor(generator): move `DetermineInputFiles`
I will need this function in the Rust+Prost codec.
